codsoft_task3/requirements.txt
torch
torchvision
transformers
pillow
requests
numpy
gradio
gTTS
deep-translator
